rename components for clarity. remove preloader
